requirements.txt
pydantic>=2.6
openai>=1.40.0
httpx>=0.27
python-dotenv>=1.0
